Add Apply method for enabling/disabling controller execution in swagger
also renamed
